// src/service-api/generated/models/callOutcomeDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * sales-crm
 * OpenAPI spec version: 1.0.0
 */

export type CallOutcomeDto = typeof CallOutcomeDto[keyof typeof CallOutcomeDto];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CallOutcomeDto = {
  NO_ANSWER: 'NO_ANSWER',
  NOT_INTERESTED: 'NOT_INTERESTED',
  INTERESTED: 'INTERESTED',
  MEETING_REQUIRED: 'MEETING_REQUIRED',
  DEAL_WON: 'DEAL_WON',
  FOLLOW_UP_LATER: 'FOLLOW_UP_LATER',
} as const;