// src/service-api/generated/models/companyStatusDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * sales-crm
 * OpenAPI spec version: 1.0.0
 */

export type CompanyStatusDto = typeof CompanyStatusDto[keyof typeof CompanyStatusDto];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CompanyStatusDto = {
  NEW: 'NEW',
  TO_CALL: 'TO_CALL',
  CALLED_NO_ANSWER: 'CALLED_NO_ANSWER',
  INTERESTED: 'INTERESTED',
  NOT_INTERESTED: 'NOT_INTERESTED',
  MEETING_REQUIRED: 'MEETING_REQUIRED',
  MEETING_SCHEDULED: 'MEETING_SCHEDULED',
  DEAL_IN_PROGRESS: 'DEAL_IN_PROGRESS',
  WON: 'WON',
  LOST: 'LOST',
  FOLLOW_UP_LATER: 'FOLLOW_UP_LATER',
} as const;